Move coverage tests to same module
