+++ indexing-service/src/test/java/org/apache/druid/indexing/compact/OverlordCompactionSchedulerTest.java
@@ -19,10 +19,8 @@
 
 package org.apache.druid.indexing.compact;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.InjectableValues;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.ImmutableMap;
 import org.apache.druid.client.indexing.ClientMSQContext;
 import org.apache.druid.guice.IndexingServiceTuningConfigModule;
 import org.apache.druid.indexer.CompactionEngine;
@@ -146,33 +144,6 @@ public class OverlordCompactionSchedulerTest
         serviceEmitter,
         OBJECT_MAPPER
     );
-  }
-
-  @Test
-  public void testCompactionSupervisorConfigSerde() throws JsonProcessingException
-  {
-    final boolean enabled = true;
-    final CompactionEngine defaultEngine = CompactionEngine.MSQ;
-    CompactionSupervisorConfig compactionSupervisorConfig =
-        OBJECT_MAPPER.readValue(
-            OBJECT_MAPPER.writeValueAsString(ImmutableMap.of("enabled", enabled, "engine", defaultEngine)),
-            CompactionSupervisorConfig.class
-        );
-    Assert.assertEquals(new CompactionSupervisorConfig(enabled, defaultEngine), compactionSupervisorConfig);
-  }
-
-  @Test
-  public void testCompactionSupervisorConfigEquality()
-  {
-    Assert.assertEquals(
-        new CompactionSupervisorConfig(true, CompactionEngine.MSQ),
-        new CompactionSupervisorConfig(true, CompactionEngine.MSQ)
-    );
-    Assert.assertNotEquals(
-        new CompactionSupervisorConfig(true, CompactionEngine.NATIVE),
-        new CompactionSupervisorConfig(true, CompactionEngine.MSQ)
-    );
-    Assert.assertNotEquals(new CompactionSupervisorConfig(true, CompactionEngine.NATIVE), "true");
   }
 
   @Test
